#1 Central datastore with collections (#9)
* Test and eslint dependencies.

* Remove initial index

* Initial datastore tests

* Remove name property from tests

It’s pretty arbitrary and I can’t see a real-world use case for it.

* Basic datastore with databases.

Straight forward, give it a name and it returns that database. Don’t
and it’ll send back the default. Creates one if none exist with that
name.

* Tests for setting the default database

* Set the default DB

* Tests for db collection.

* Fix tests

* Create the collection component

To be composed into the database through composition!

* Add collection to db

Also, fix up default

* Install aproba and alter main file

* Fix to travis.