v0.2.3 (#42)
* disable event emission

* events tests

* swapped events component

It completely relies on being attached to a composed object. That
object then has it’s own responsibility for creating and managing
events.

* better events tests

* remove events from database

* remove meta field

In place of just name, or whatever. No need for the complication when
data is in it’s own property anyway.

* remove events from collection

* remove memOnly option from collection

* more memOnly

* renaming for clarity

database.data -> database.collections
collection[] -> collection.data[]

* update rx to 5.x

Sad I didn’t notice that before, was having some weird behavior.

* document events

* fix up collection test

* collection-created event

* needs to send the name

* database-created event

* v0.2.3

closes #30 
closes #34